modified:   Monitor.cpp 	modified:   Monitor.h 	modified:   pins.h 	modified:   station.ino
	modified:   src/station/Monitor.cpp
	modified:   src/station/Monitor.h
	modified:   src/station/pins.h
	modified:   src/station/station.ino

+++ src/station/Monitor.h
@@ -22,7 +22,7 @@ class BatteryMonitor {
 
 public:
   int get_adc_value();
-  int get_adc_voltage();
+  float get_adc_voltage();
   float get_voltage(int adc_pin);
   float get_devider_current();
